caesar cipher algorithm
Python script to encrypt or decrypt a text.
This program implements the caesar cipher algorithm which helps in encrypting and decrypting the text code.
The program prompts the user to encrypt or decrypt by asking him to provide input in the form of a  text.
And the corresponding input text is encrypted or decrypted according to the chosen operation i.e encryption or decryption.
Note that the program asks for the shift value. Since, we are implementing the caesar cipher algorithm, to encrypt or to decrypt the shift value will be 3.(if you want you can use any number less than 25 as a shift value).Why i am specifying as 25 because the indexing starts from 0 and there are totally 26 alphabets in english.